rename : 폴더 이동 (Kiosk_Lv6 -> Kiosk_Lv7)
