+++ src/kiosk/Kiosk_Lv7/Cart.java
@@ -1,6 +1,5 @@
-package kiosk.Kiosk_Lv6;
+package kiosk.Kiosk_Lv7;
 
-import java.util.ArrayList;
 import java.util.List;
 
 // 장바구니 관리
