+++ src/kiosk/Kiosk_Lv7/Kiosk.java
@@ -1,4 +1,4 @@
-package kiosk.Kiosk_Lv6;
+package kiosk.Kiosk_Lv7;
 
 import java.util.ArrayList;
 import java.util.List;
