+++ src/kiosk/Kiosk_Lv7/Menu.java
@@ -1,4 +1,4 @@
-package kiosk.Kiosk_Lv6;
+package kiosk.Kiosk_Lv7;
 
 import java.util.List;
